Refactor company responsibilities into dedicated services
