+++ game/models/company/constants.ts
@@ -1,0 +1,7 @@
+export const XP_PER_LEVEL = 100;
+export const TASK_XP_REWARD = 10;
+export const ENERGY_COST_PER_TICK_WORKING = 10.0;
+export const ENERGY_REGEN_PER_TICK_RESTING = 10; // Resting in breakroom
+export const IDLE_ENERGY_REGEN_PER_TICK = 2.5;
+export const ENERGY_REST_THRESHOLD = 20;
+export const OFF_DUTY_DURATION_TICKS = 16;
